requirements_v2.txt
# Core
streamlit==1.32.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pypdf==4.2.0
openai==1.25.2
python-dotenv==1.0.1
numpy==1.26.4

# Hybrid search
rank-bm25==0.2.2

# Evaluation
scikit-learn==1.4.1.post1